+++ test/models/commits_test.rb
@@ -35,6 +35,33 @@ module Shipit
       end
     end
 
+    test '.create_from_github handle commits with empty message' do
+      assert_difference -> { Commit.count }, +1 do
+        @stack.commits.create_from_github!(
+          resource(
+            sha: '2adaad1ad30c235d3a6e7981dfc1742f7ecb1e85',
+            author: {},
+            committer: {},
+            commit: {
+              author: {
+                name: 'Lando Walrussian',
+                email: 'walrus@shopify.com',
+                date: Time.now,
+              },
+              committer: {
+                name: 'Lando Walrussian',
+                email: 'walrus@shopify.com',
+                date: Time.now,
+              },
+              message: '',
+            },
+          ),
+        )
+      end
+      commit = Commit.last
+      refute_predicate commit, :revert?
+    end
+
     test '.create_from_github handle PRs merged by another Shipit stacks' do
       assert_difference -> { Commit.count }, +1 do
         @stack.commits.create_from_github!(